Support different types for the index buffer.
Using u32 all the time is massively wasteful for many workloads,
since we rarely have meshes with more than 65535 verticies (so we can get away
with u16s).  GL also supports u8 index buffers, which are handy for very small
meshes so we add support for those as well since it's easy enough to do.