Refactor TextureUnits to create ResourceMapper.
`ResourceMapper` is a way to map resource handles with bindings. The
allocation scheme is the same as the old texture binding scheme: we try
to exhaust the hardware limit (passed as an input to `ResourceMapper`),
and when we cannot get a binding anymore, we start reusing idlding ones.

We mark a binding as idle once we drop the “binding handle” (for
instance, `InUseTexture` for textures, and probably `InUseUniBuffer` for
uniform buffers). We unmark a binding once we start reusing it.